% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civil_data_dict}
\alias{civil_data_dict}
\title{AOC VCAP data layout}
\format{
A data frame with 852 rows and 10 variables:
\describe{
\item{table_type}{description}
\item{table_code}{description}
\item{table_id}{description}
\item{col_names}{description}
\item{type}{description}
\item{null_option}{description}
\item{description}{description}
\item{start}{description}
\item{end}{description}
\item{length}{description}
}
}
\source{
\url{https://www.nccourts.gov/services/remote-public-access-program/rpa-resources}
}
\usage{
civil_data_dict
}
\description{
A dataset containing the layout of tables in the AOC civil court extract
}
\keyword{datasets}
