% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatCode.R
\name{getCatCode}
\alias{getCatCode}
\title{Get AOC category}
\usage{
getCatCode(record_category)
}
\arguments{
\item{record_category}{plain language record category in short or long form}
}
\value{
string with corresponding category code
}
\description{
Simple utility function to convert a plain language table category code to
the corresponding file name per AOC naming convention.
}
\examples{
getCatCode('c')
}
