% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonLayout.R
\name{getCommonLayout}
\alias{getCommonLayout}
\title{Retrieve the common layout for a table}
\usage{
getCommonLayout(cat_code, data_dict, add_line = 1)
}
\arguments{
\item{cat_code}{category code for the record ('NOBC', etc.)}

\item{data_dict}{a dataframe of data dictionary values with cols start, end & col_names}

\item{add_line}{a flag to specify a generic catch-all column before initial splitting (default is 1)}
}
\value{
a dataframe with layout specs
}
\description{
Simple utility function to convert a record category code into a dataframe
with the common record layout based on a specified data dictionary dataframe.
}
\examples{
cat_code <- getCatCode('c')
getCommonLayout(cat_code, civil_data_dict)
}
