% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{importFiles}
\alias{importFiles}
\title{Import civil data}
\usage{
importFiles(record_category = "c", data_dict, source_dir)
}
\arguments{
\item{record_category}{Define type of record. Default is 'case'.
'case' or 'c' for case record,
'abstract' or 'a' for abstract record
'support' or 's' for support record}

\item{data_dict}{the built-in data dictionary}

\item{source_dir}{the directory containing files from AOC, with filenames NOBC0001 etc.}
}
\value{
a very large dataframe
}
\description{
Import function to load in all tables exported by AOC of a specified record
category, using a predefined data dictionary of start and end positions. Produces
a dataframe with seven columns, where the first six are the the common layout
for the specified record category, followed by the fixed width line.
\cr\cr NOTE: Embedded nulls in some files required advanced cleaning of files using perl:
\cr\cr\verb{$ perl -pe 's/\\0/ /g' < NOBC0001 > NOBC0001fix}
\cr\cr Run time: 2 minutes/MTD
}
\examples{
importFiles('c', civil_data_dict, 'vcap/')
}
