% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllSpecs.R
\name{load_all_specs}
\alias{load_all_specs}
\title{Use a data dictionary to load specs}
\usage{
load_all_specs(civil_data_dict)
}
\arguments{
\item{civil_data_dict}{A dataframe describing the data}
}
\description{
Use a data dictionary to load specs
}
