% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllSpecs.R
\name{loadAllSpecs}
\alias{loadAllSpecs}
\title{Use a data dictionary to load specs}
\usage{
loadAllSpecs(civil_data_dict)
}
\arguments{
\item{civil_data_dict}{A dataframe describing the data}
}
\value{
A list of data layouts named for each table
}
\description{
Use a data dictionary to load specs
}
\examples{
#load all specifications per a specified data dictionary
all_specs <- loadAllSpecs(civil_data_dict)
}
