% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCivilTable.R
\name{loadCivilTable}
\alias{loadCivilTable}
\title{Use table specs to load in a civil data table}
\usage{
loadCivilTable(table_path, all_specs)
}
\arguments{
\item{table_path}{Description}

\item{all_specs}{Description}
}
\value{
A parsed table
}
\description{
Use table specs to load in a civil data table
}
\examples{
#' #load all specifications per a specified data dictionary
all_specs <- loadAllSpecs(civil_data_dict)

#load relevant tables from civil data
issue_type <- loadCivilTable('../../aoc/2022/sep/NOBC23', all_specs)
party_names <- loadCivilTable('../../aoc/2022/sep/NOBC48', all_specs)
}
