% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpec.R
\name{loadSpec}
\alias{loadSpec}
\title{Subprocess for vectorizing the load of individual specs for a table}
\usage{
loadSpec(table)
}
\arguments{
\item{table}{A table}
}
\value{
A table of column specs
}
\description{
Subprocess for vectorizing the load of individual specs for a table
}
