% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCivilData.R
\name{parseCivilData}
\alias{parseCivilData}
\title{Parse civil data}
\usage{
parseCivilData(record_category = "c", data_dict, source_dir, record_id = NA)
}
\arguments{
\item{record_category}{Define type of record. Default is 'case'.
\cr\cr 'case' or 'c' for case record,
\cr\cr 'abstract' or 'a' for abstract record
\cr\cr 'support' or 's' for support record}

\item{data_dict}{preloaded data dictionary}

\item{source_dir}{source directory for the separated files}

\item{record_id}{(optional) provide the record type number if you only want one table}
}
\value{
list of dataframes
}
\description{
Function that accepts info about the files you want to parse and reads them
into a list of dataframes using a predefined data dictionary.
\cr\cr Run time: 18 seconds for largest tables, 3 minutes for all/MTD
}
\examples{
#don't run these
#parseCivilData('c', civil_data_dict, 'sep/', record_id = 48) #specific
#parseCivilData('c', civil_data_dict, 'sep/') #all
}
