% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCivilDirectory.R
\name{processCivilDirectory}
\alias{processCivilDirectory}
\title{Process a directory of raw civil extract files}
\usage{
processCivilDirectory(
  source_dir,
  dest_dir,
  record_category = "c",
  data_dict = vcapr::civil_data_dict
)
}
\arguments{
\item{source_dir}{directory containing raw files}

\item{dest_dir}{directory for writing files}

\item{record_category}{define type of record. Default is 'case'
\cr\cr'case' or 'c' for case record,
\cr\cr'abstract' or 'a' for abstract record
\cr\cr'support' or 's' for support record}

\item{data_dict}{layout file containing the data dictionary. Default is built-in
data dictionary}
}
\value{
a dataframe reporting out rows details and parse time
}
\description{
Import, parse and write an entire directory of extract data provided by the
N.C. Administrative Office of the Courts for the VCAP system managing civil courts.
By default, parsing follows the VCAP extract file layout for case tables.
}
\examples{
process_results <- processCivilDirectory('/vcap/tables/', 'vcap/2022/sep_test')
}
