% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCivilFile.R
\name{processCivilFile}
\alias{processCivilFile}
\title{Process a raw civil extract file}
\usage{
processCivilFile(
  file_path,
  dest_dir,
  record_category = "c",
  data_dict = vcapr::civil_data_dict
)
}
\arguments{
\item{file_path}{path containing raw file with filename NOBC0001 or NOBC0002 etc.}

\item{dest_dir}{directory for writing files}

\item{record_category}{define type of record. Default is 'case'
\cr\cr'case' or 'c' for case record,
\cr\cr'abstract' or 'a' for abstract record
\cr\cr'support' or 's' for support record}

\item{data_dict}{layout file containing the data dictionary. Default is built-in
data dictionary}
}
\value{
a dataframe reporting out rows details and parse time
}
\description{
Import, parse and write single raw file from the extract data provided by the
N.C. Administrative Office of the Courts for the VCAP system managing civil courts.
By default, parsing follows the VCAP extract file layout for case tables.
}
\note{
Because this function is often used to loop through multiple files, it currently
appends data to existing files with the same table name in your destination directory.
Be careful not to double-count your data!
}
\examples{
processCivilFile('/vcap/NOBC0001', 'vcap/2022/sep_test')
}
