% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSepTables.R
\name{writeSepTables}
\alias{writeSepTables}
\title{Write tables}
\usage{
writeSepTables(table_list, dest_dir, record_category = "c", threads = 8)
}
\arguments{
\item{table_list}{a list of tables contained separated data from AOC civil extract}

\item{dest_dir}{destination directory for the resulting separated files}

\item{record_category}{Define type of record. Default is 'case'.
\cr\cr 'case' or 'c' for case record,
\cr\cr 'abstract' or 'a' for abstract record
\cr\cr support' or 's' for support record}

\item{threads}{the number of threads for vroom to use. Default is 8.}
}
\value{
No return value
}
\description{
Function that accepts a list of split tables stored in a list object and
writes to a fixed width file in a specified directory. Tables can be written
individually or all at once.
\cr\cr NOTE: This is the slowest step in the process, and seems unnecessary
given that the data is already loaded. Is there a way to reparse the data,
instead of writing and loading separately?
\cr\cr Run time: ~ 4 hours for all, ~15 min for larger tables/MTD
}
\examples{
#don't run these, very slow
#writeSepTables(case_tables, 'sep_test/', 'c', threads = 8) #specific table
#writeSepTables(case_tables['03'], 'sep_test/', 'c', threads = 8) #all tables
}
