% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCivilTable.R
\name{load_civil_table}
\alias{load_civil_table}
\title{Use table specs to load in a civil data table}
\usage{
load_civil_table(table_path, all_specs)
}
\arguments{
\item{all_specs}{}
}
\description{
Use table specs to load in a civil data table
}
