% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTables.R
\name{splitTables}
\alias{splitTables}
\title{Split civil data tables}
\usage{
splitTables(combined_tables)
}
\arguments{
\item{combined_tables}{a dataframe containing the combined tables from AOC}
}
\value{
large list
}
\description{
Function that accepts a dataframe of combined tables from the AOC civil extract
and splits them into component parts based on the record type. Output is a large
list that we can later use to write indidividual separated files.
\cr\cr Run time: ~ 8 minutes/MTD
}
\examples{
all_tables <- importFiles('c', civil_data_dict, 'vcap/')
splitTables(all_tables)
}
